requirements.txt
requests
coverage
ddt
xmlrunner
requests-ntlm
redis
dohq-tfs
cassandra-driver